bash branch completion decolorization fix
git branch really isn't meant to be used in scripts -- The git porcelain functions are. If you do use git branch in a script, but sure to disable color, or some systems will complete branches named 'foo' as 'foo^[[m'. Also be prepared that changes to git branch output in the future may break completion.